requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
